[BE] Removed unused Controller & Add AWS Secrets Manager
