--- src/main/java/com/mainlab/model/ContentType.java
@@ -1,13 +1,0 @@
-package com.mainlab.model;
-
-import lombok.Getter;
-
-@Getter
-public enum ContentType {
-    PROFILE,
-    LAB,
-    ALGORITHM,
-    ARCHITECTURE,
-    PROJECT,
-    ;
-}
