--- src/main/resources/sqlmap/ContentSql.xml
@@ -1,46 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="ContentSql">
-    <select id="selectContentList" resultType="com.mainlab.model.Content">
-        SELECT
-            `content_id` as `contentId`,
-            `content_type` as `contentType`
-        FROM
-            `content`
-    </select>
-    <select id="selectContent" parameterType="int" resultType="com.mainlab.model.Content">
-        SELECT
-            `content_id` as `contentId`,
-            `content_type` as `contentType`
-        FROM
-            `content`
-        WHERE
-            `content_id` = #{contentId}
-    </select>
-    <select id="selectContentListByContentType" parameterType="map" resultType="com.mainlab.model.Content">
-        SELECT
-            `content_id` as `contentId`,
-            `content_type` as `contentType`
-        FROM
-            `content`
-        WHERE
-            `content_type` = #{contentType}
-    </select>
-
-    <insert id="insertContent" parameterType="com.mainlab.model.Content" useGeneratedKeys="true">
-        INSERT INTO `content`
-        (
-            `content_type`,
-            `created`,
-            `modified`
-        ) VALUES
-        (
-            #{contentType},
-            NOW(),
-            NOW()
-        )
-        <selectKey keyProperty="contentId" resultType="int">
-            SELECT LAST_INSERT_ID();
-        </selectKey>
-    </insert>
-</mapper>
